package-lock.jsonの削除とpnpm-lock.yamlの修正
